apps/web/requirements.txt
asgiref==3.6.0 ; python_version >= "3.9" and python_version < "4.0"
django==3.2.18 ; python_version >= "3.9" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.9" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.9" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.9" and python_version < "4.0"
setuptools==67.4.0 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
whitenoise==6.4.0 ; python_version >= "3.9" and python_version < "4.0"
dj-database-url==2.1.0
djangorestframework==3.14.0
typing_extensions==4.10.0
gunicorn
django-cors-headers==3.3.0
